[INTERNAL] Return type of GetLastError is DWORD
Fixes warning issued by clang/c2 3.7 on windows:
comparison of integers of different signs: 'const unsigned long' and 'const long'

+++ include/seqan/system/system_condition.h
@@ -111,7 +111,7 @@ waitFor(Condition &cond, long timeoutMilliSec, bool & inProgress)
 #ifdef STDLIB_VS
     inProgress = (SleepConditionVariableCS(&cond.data_cond, &cond.csPtr->data_cs, timeoutMilliSec) == 0);
     if (inProgress)
-        SEQAN_ASSERT_EQ(GetLastError(), ERROR_TIMEOUT);
+        SEQAN_ASSERT_EQ(GetLastError(), static_cast<DWORD>(ERROR_TIMEOUT));
 #else
     if (timeoutMilliSec != Condition::Infinite)
     {
